latest release upload

+++ b_process.php
@@ -16,17 +16,28 @@
 <?php
     if($_SERVER["REQUEST_METHOD"] == "POST"){
         $table_name = $_POST["op_table"];
-        $card_number = $_POST["card_number"];
-        $start_time = $_POST["start_time"];
-        $query_to_delete = "delete from $table_name where card_id='".$card_number."' and start_time='".$start_time."';";
-        $delete_result = $conn->query($query_to_delete);
-        if(!$delete_result){
-            die($delete_result);
+        if($table_name=='bookingsGroceries' || $table_name=='bookingsLiquor') {
+          $card_number = $_POST["card_number"];
+          $start_time = $_POST["start_time"];
+          $query_to_delete = "delete from $table_name where card_id='".$card_number."' and start_time='".$start_time."';";
+          $delete_result = $conn->query($query_to_delete);
+          if(!$delete_result){
+              die($delete_result);
         }
         $message = "Booking of card number $card_number at $start_time cancelled";
-        
 
+       } else if ($table_name=='bookingsgroceriesliquor') {
+          $card_number = $_POST["card_number1"];
+          $card_number1 = $_POST["card_number2"];
+          $start_time = $_POST["start_time"];
+          $query_to_delete = "delete from $table_name where card_id_groceries='".$card_number."' and card_id_liquor='".$card_number1."' and start_time='".$start_time."';";
+          $delete_result = $conn->query($query_to_delete);
+          if(!$delete_result){
+              die($delete_result);
+        }
+        $message = "Booking of card number $card_number and card number $card_number1 at $start_time cancelled";
 
+        }        
 
 
     }
@@ -69,7 +80,7 @@
 </body>
 <!-- Footer -->
 <footer class="page-footer font-small blue">
-  <div class="footer-copyright text-center py-2">© 2020 Copyright:
+  <div class="footer-copyright text-center py-2">© 2021 Copyright:
     <a href="acknowledgement.php">Team Page</a>
   </div>
 </footer>
